fixed bad setinterval clears on front and back

+++ backend/src/socket/socketEvents.ts
@@ -543,8 +543,7 @@ export class SocketEvents {
         const size = this.server.sockets.adapter.rooms.get(key)?.size;
         if (size === 1) {
           const room = await this.roomsService.getRoom(key);
-          if (room)
-            this.roomsService.deleteRoom(key);
+          if (room) this.roomsService.deleteRoom(key);
         }
       });
     });
@@ -607,20 +606,15 @@ export class SocketEvents {
   }
 
   async endGame(game: IGame, winnerSocketId: string, loserSocketId: string) {
-    console.log(`winenrsocketid: ${winnerSocketId}, loserSocketId: ${loserSocketId}`)
     const winnerId = this.mapUserInGame.get(winnerSocketId);
     const loserId = this.mapUserInGame.get(loserSocketId);
     let winner = "";
     let loser = "";
-    console.log("END GAMEEEEEEEEEEEEEEEEEEEEEE")
-    console.log(`winnerId: ${winnerId}, loserId: ${loserId}`);
     if (winnerId && loserId) {
       await this.userService.findUsersById(winnerId).then((user) => {
-        console.log("winner user: ", user)
         if (user) winner = user.username;
       });
       await this.userService.findUsersById(loserId).then((user) => {
-        console.log("looser user: ", user)
         if (user) loser = user.username;
       });
       await this.userService.updateHistory(
@@ -633,6 +627,7 @@ export class SocketEvents {
       await this.userService.updateAchive(loserId);
     }
     this.server.to(game.id).emit("end_game", { winner: winner, loser: loser });
+    console.log("clear interval: " + game.intervalId);
     clearInterval(game.intervalId);
     games = games.filter((g) => g.id !== game.id);
     this.mapUserInGame.delete(winnerSocketId);
@@ -659,7 +654,7 @@ export class SocketEvents {
       } else {
         this.endGame(game, game.player1.socketId, game.player2.socketId);
       }
-      games = games.filter(g => g.id !== game?.id);
+      games = games.filter((g) => g.id !== game?.id);
     }
     this.mapUserInGame.delete(client.id);
   }
@@ -701,24 +696,26 @@ export class SocketEvents {
       for (let [key, value] of this.mapUserInGame.entries()) {
         if (client.id === key) {
           const userDb = await this.userService.findUsersById(Number(value));
-          console.log(userDb)
-          this.server.to(data.roomId).emit("user_leave_room", { username: userDb?.username });
+          console.log(userDb);
+          this.server
+            .to(data.roomId)
+            .emit("user_leave_room", { username: userDb?.username });
         }
       }
-      console.log("J AI LEAVE LA GAME DEPUIS ON LEAVE GAME")
+      console.log("J AI LEAVE LA GAME DEPUIS ON LEAVE GAME");
       client.leave(data.roomId);
       //this.mapUserInGame.delete(client.id);
-      
+
       const game = this.findGameByConnectedSocket(client.id);
       if (game) {
-        console.log()
-        console.log(game)
+        console.log();
+        console.log(game);
         if (game.player1.socketId === client.id) {
           this.endGame(game, game.player2.socketId, game.player1.socketId);
         } else {
           this.endGame(game, game.player1.socketId, game.player2.socketId);
         }
-        games = games.filter(g => g.id !== game?.id);
+        games = games.filter((g) => g.id !== game?.id);
       }
       this.mapUserInGame.delete(client.id);
       const nbClient = this.server.sockets.adapter.rooms.get(data.roomId)?.size;
@@ -758,7 +755,9 @@ export class SocketEvents {
     });*/
     for (let [key, value] of this.mapUserInGame.entries()) {
       if (value === userId) {
-        this.server.to(client.id).emit("join_game_error", { error: "You are already in a party" });
+        this.server
+          .to(client.id)
+          .emit("join_game_error", { error: "You are already in a party" });
         return;
       }
     }
@@ -773,25 +772,32 @@ export class SocketEvents {
     const room = await this.roomsService.findRoomById(data.roomId);
     //j'enleve ca car le find fonctionne mal, il cherche sur toutes les rooms chat compris,
     //si c pour trouver si deja en partie faut modifier
-    console.log(this.mapUserInGame)
+    console.log(this.mapUserInGame);
 
     //check if user is in a party
 
-    if (/*socketRooms.length > 0
-      || */(connectedSockets && connectedSockets?.size > 1)) {
-      this.server.to(client.id).emit("join_game_error", { error: "Room is full" });
+    if (
+      /*socketRooms.length > 0
+      || */ connectedSockets &&
+      connectedSockets?.size > 1
+    ) {
+      this.server
+        .to(client.id)
+        .emit("join_game_error", { error: "Room is full" });
       return;
     } else {
-      console.log(typeof data.roomId)
+      console.log(typeof data.roomId);
       await client.join(data.roomId);
-      this.roomsService.updateRoomReady(data.roomId, false, true, true)
+      this.roomsService.updateRoomReady(data.roomId, false, true, true);
       this.mapUserInGame.set(client.id, userId);
       if (room && room.private === true) {
         const result: boolean = this.checkIfUserFound(room, client.id);
         //console.log("result: " + result)
         //console.log(this.server.sockets.adapter.rooms.get(data.roomId))
         if (result === false) {
-          this.server.to(client.id).emit("join_game_error", { error: "You are spectactor" });
+          this.server
+            .to(client.id)
+            .emit("join_game_error", { error: "You are spectactor" });
           return;
         }
       }
@@ -807,7 +813,7 @@ export class SocketEvents {
         username: user.username,
         nbClient: nbClient
       });*/
-      console.log("--------")
+      console.log("--------");
       console.log(this.server.sockets.adapter.rooms.get(data.roomId));
       console.log(this.mapUserInGame);
       client.to(data.roomId).emit("updateUserRdy");
@@ -815,16 +821,16 @@ export class SocketEvents {
       //const nbClient = this.server.sockets.adapter.rooms.get(data.roomId)?.size;
       let i: number = 1;
       loop?.forEach((key) => {
-        console.log(key)
+        console.log(key);
         this.mapUserInGame.forEach(async (value2, key2) => {
           if (key === key2) {
             console.log("FOUND");
             const userDb = await this.userService.findUsersById(Number(value2));
-            console.log(userDb)
+            console.log(userDb);
             this.server.to(data.roomId).emit("join_game_success", {
               roomId: data.roomId,
               username: userDb?.username,
-              nbClient: i
+              nbClient: i,
             });
             ++i;
           }
@@ -897,41 +903,49 @@ export class SocketEvents {
 
   @UseGuards(JwtGuard)
   @SubscribeMessage("updateTypeGame")
-  updateTypeGame(@MessageBody() data: UpdateTypeRoom, @ConnectedSocket() client: Socket) {
+  updateTypeGame(
+    @MessageBody() data: UpdateTypeRoom,
+    @ConnectedSocket() client: Socket
+  ) {
     //client.to().emit()
-    console.log("data")
-    console.log(data)
-    client.to(data.roomId).emit("updateTypeGameFromServer", { type: data.type });
+    console.log("data");
+    console.log(data);
+    client
+      .to(data.roomId)
+      .emit("updateTypeGameFromServer", { type: data.type });
   }
 
   @UseGuards(JwtGuard)
   @SubscribeMessage("userIsRdy")
-  async gameIsRdy(@MessageBody() data: UserIdRdy, @ConnectedSocket() client: any) {
+  async gameIsRdy(
+    @MessageBody() data: UserIdRdy,
+    @ConnectedSocket() client: any
+  ) {
     const user: TokenUser = client.user;
     const connectedSockets = this.server.sockets.adapter.rooms.get(data.uid);
 
-    console.log("data rdy")
-    console.log(data)
-    console.log("user")
-    console.log(user)
+    console.log("data rdy");
+    console.log(data);
+    console.log("user");
+    console.log(user);
     if (user.username === data.usr1)
       await this.roomsService.updateRoomReady(data.uid, data.rdy, true, false);
     else if (user.username === data.usr2)
       await this.roomsService.updateRoomReady(data.uid, data.rdy, false, true);
     //when two user are connected, and both are rdy, game must start
     const getRoom = await this.roomsService.getRoom(data.uid);
-    if (connectedSockets?.size === 2
-      && getRoom?.player_one_rdy === true
-      && getRoom.player_two_rdy === true) {
+    if (
+      connectedSockets?.size === 2 &&
+      getRoom?.player_one_rdy === true &&
+      getRoom.player_two_rdy === true
+    ) {
       // console.log(connectedSockets)
       //let socket2 = connectedSockets?.values().next().value;
       let socket2: string | undefined = undefined;
       connectedSockets?.forEach((key) => {
-        if (key !== client.id)
-          socket2 = key;
+        if (key !== client.id) socket2 = key;
       });
-      if (socket2 === undefined)
-        return ;
+      if (socket2 === undefined) return;
       let newGame = new Game(data.uid, client.id, socket2, 11);
       let powerUps = newGame.powerUps;
       let player1 = newGame.player1;
@@ -941,19 +955,19 @@ export class SocketEvents {
       console.log("Starting game");
       client.emit("start_game", { side: 1 });
       client.to(data.uid).emit("start_game", { side: 2 });
-      setInterval(() => {
+      newGame.intervalId = setInterval(() => {
         this.update(newGame);
         client.emit("on_game_update", {
           player1,
           player2,
           ball,
-          powerUps
+          powerUps,
         });
         client.to(data.uid).to(data.uid).emit("on_game_update", {
           player1,
           player2,
           ball,
-          powerUps
+          powerUps,
         });
       }, 1000 / FPS);
     }
